+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.16.6
 
-- New: Add support for passing options to `ShopifyAPI.Bulk.Query.exec/3`, with `group_objects`
+- New: Add support for passing options to `ShopifyAPI.Bulk.Query.exec!/3`, with `group_objects`
   as a new option to group the returned objects by a specified key.
 
 ## 0.16.5